Fail gracefully if the nategosearch type doesn't exist
git-svn-id: file:///Users/gauthierm/svnrepo/public/store/trunk@30641 94490950-a86c-4493-9ccb-530f72933e47

+++ Store/admin/components/Category/Edit.php
@@ -149,6 +149,9 @@ class StoreCategoryEdit extends AdminDBEdit
 	{
 		$type = NateGoSearch::getDocumentType($this->app->db, 'category');
 
+		if ($type === null)
+			return;
+
 		$sql = sprintf('delete from NateGoSearchQueue
 			where document_id = %s and document_type = %s',
 			$this->app->db->quote($this->id, 'integer'),
